Extract the copperline library crate
All emulator modules now live in a library crate (src/lib.rs) with the
copperline binary reduced to a thin CLI wrapper. This lets alternative
frontends, fuzz targets, and test harnesses link the deterministic core
directly instead of going through the binary.

Moves that give the library a complete surface:
- build_machine moves from main.rs to emulator::build_machine (the
  configuration screen's Run button already needed it from the library
  side).
- The env-override resolvers (resolve_bundled_rom, resolve_overscan,
  resolve_pixel_aspect, resolve_phosphor) and about_machine_lines move
  from main.rs to config.rs.
- RawConfig, Config::load_raw, and App::open_launcher widen to pub for
  the binary; Default impls and an is_empty accompany the public new()
  and len() methods clippy now lints on the library surface.
