+++ src/inputsched.rs
@@ -137,6 +137,11 @@ impl ReplayInputLog {
     pub fn len(&self) -> usize {
         self.events.len()
     }
+
+    #[cfg(test)]
+    pub fn is_empty(&self) -> bool {
+        self.events.is_empty()
+    }
 }
 
 #[cfg(test)]
